ignore incomplete test about the physics converter
